Updated weather fetcher to save tts file. Updated constants to have weather subdirectory. Added requirements.txt
Weather fetcher now calls gTTS to generate text to speech file and saves to the weather subfolder in the media directory.

Added requirements.txt to track required packages. As the project goes more complex, this will be nice so we can keep track of everything we need to run the app.